Move swap leg amount to amount package
Better fit for class
