+++ modules/market/src/test/java/com/opengamma/strata/market/amount/LegAmountsTest.java
@@ -19,7 +19,6 @@ import com.google.common.collect.ImmutableList;
 import com.opengamma.strata.basics.currency.Currency;
 import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.basics.currency.FxRate;
-import com.opengamma.strata.market.product.swap.SwapLegAmount;
 import com.opengamma.strata.product.common.PayReceive;
 import com.opengamma.strata.product.swap.SwapLegType;
